64-bit changes
SCCS-vsn: sys/netiso/iso.c 8.3
SCCS-vsn: sys/netiso/iso_snpac.c 8.2
SCCS-vsn: sys/netiso/if_eon.c 8.2

+++ usr/src/sys/netiso/if_eon.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)if_eon.c	8.1 (Berkeley) %G%
+ *	@(#)if_eon.c	8.2 (Berkeley) %G%
  */
 
 /***********************************************************
@@ -152,7 +152,7 @@ eonattach()
  */
 eonioctl(ifp, cmd, data)
 	register struct ifnet *ifp;
-	int cmd;
+	u_long cmd;
 	register caddr_t data;
 {
 	int s = splimp();
