+++ usr/src/sys/netiso/iso.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)iso.c	8.2 (Berkeley) %G%
+ *	@(#)iso.c	8.3 (Berkeley) %G%
  */
 
 /***********************************************************
@@ -387,7 +387,7 @@ caddr_t			buf;		/* RESULT: network portion of address here */
 /* ARGSUSED */
 iso_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
